files/requirements.txt
PyQt5>=5.15
opencv-python-headless
face_recognition
mutagen
python-magic
pillow
numpy
scikit-learn
python-docx
pytesseract
pdfminer.six
python-barcode
pyzbar
pandas
nltk
requests
pycryptodome
pyyaml
tqdm
soundfile
SpeechRecognition
beautifulsoup4